added obstacle, and scale to vehicle
also added some small fixes

+++ AAI-assignment/behaviour/AllignmentBehaviour.cs
@@ -36,7 +36,7 @@ namespace AAI_assignment.behaviour
 
             if (NeighbourCount > 0)
             {
-                averageHeading /= NeighbourCount;
+                averageHeading /= (double)NeighbourCount;
                 averageHeading -= ME.Heading;
             }
